+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 5.0.0 (Unreleased)
+* Support PHP 7.4 typed properties
 * [Classlike, function and constant completions now show their FQSEN in the `detail` property instead of the `label`](https://gitlab.com/Serenata/Serenata/issues/269)
 * Fix namespace autocompletion suggestions not having their backslashes properly escaped
 * Fix classlike autocompletion suggestions not having their backslashes properly escaped when starting with a leading slash